correção

+++ aulapratica7_lista7_codigo6_avaliativa.c
@@ -1,5 +1,5 @@
 #include <stdio.h>
-int soma(int n){
+float soma(int n){
     int soma=0;
     for(int i = 1; i<=n; i++){
         soma += 1/i;
@@ -10,6 +10,6 @@ int main(){
     int num;
     printf("Digite um numero inteiro positivo: ");
     scanf("%d", &num);
-    printf("A soma eh: %d\n", soma(num));
+    printf("A soma eh: %.2f\n", soma(num));
     return 0;
 }
